Add comments clarifying loops and migrate exports
