+++ tests/test_exports.py
@@ -3,7 +3,7 @@ import pathlib
 import pandas as pd
 
 sys.path.insert(0, str(pathlib.Path(__file__).resolve().parents[1]))
-from exports import export_data
+from trend_analysis.export import export_data
 
 
 def test_export_data(tmp_path):
